Reduce memory use
git-svn-id: svn://svn.code.sf.net/p/tass64/code/trunk@2077 111d7451-93fa-4e81-90d4-8039f76616e9

+++ dictobj.h
@@ -33,7 +33,7 @@ typedef struct Dict {
     size_t len, max;
     struct pair_s *data;
     struct pair_s val[1];
-    size_t idx[1];
+    uint8_t idx[1];
     Obj *def;
 } Dict;
